refactor: avoid using utils
