+++ lib/Server.js
@@ -10,7 +10,6 @@ const killable = require('killable');
 const express = require('express');
 const { validate } = require('schema-utils');
 const normalizeOptions = require('./utils/normalizeOptions');
-const colors = require('./utils/colors');
 const routes = require('./utils/routes');
 const getSocketServerImplementation = require('./utils/getSocketServerImplementation');
 const getCompilerConfigArray = require('./utils/getCompilerConfigArray');
@@ -829,6 +828,26 @@ class Server {
 
       this.logger.info('Project is running at:');
 
+      // TODO change it on https://www.npmjs.com/package/colorette
+      const colors = {
+        info(useColor, msg) {
+          if (useColor) {
+            // Make text blue and bold, so it *pops*
+            return `\u001b[1m\u001b[34m${msg}\u001b[39m\u001b[22m`;
+          }
+
+          return msg;
+        },
+        error(useColor, msg) {
+          if (useColor) {
+            // Make text red and bold, so it *pops*
+            return `\u001b[1m\u001b[31m${msg}\u001b[39m\u001b[22m`;
+          }
+
+          return msg;
+        },
+      };
+      
       if (server) {
         this.logger.info(`Server: ${colors.info(useColor, server)}`);
       }
