--- lib/utils/findPort.js
@@ -1,6 +1,0 @@
-'use strict';
-
-// TODO(@anshumanv) - Drop in next major release
-const getFreePort = require('../Server').getFreePort;
-
-module.exports = getFreePort;
